refactor: CourseApplicationService를 Command와 Query 서비스로 분리
